move Roslin operator tests to test_roslin_operator
